SearchEngineApplication/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai>=1.108.0
pydantic>=2.8.0,<3.0.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx>=0.26.0,<0.29.0
weaviate-client==4.16.9